I tried fixing delete

+++ uek295-1-LeIs-Author/src/main/java/ch/noseryoung/uek2951LeIsAuthor/domain/author/AuthorService.java
@@ -1,6 +1,7 @@
 package ch.noseryoung.uek2951LeIsAuthor.domain.author;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import javax.management.InstanceAlreadyExistsException;
@@ -30,7 +31,11 @@ public class AuthorService {
        return repository.save(author);
     }
 
-    public void deleteAuthor(Integer authorID) throws AuthorController.AuthorNotFoundException, NoSuchElementException {
+    public void deleteAuthor(Integer authorID) throws AuthorController.AuthorNotFoundException, NoSuchElementException, InstanceNotFoundException {
+       if (!repository.existsById(authorID)) {
+           throw new InstanceNotFoundException("The id " + authorID + " has been deleted");
+        }
+
         repository.deleteById(authorID);
     }
